feat(RoomescapeRepository): Repository 인터페이스 도입 및 구현체 분리
